Input validation

+++ src/main/java/cz/svjis/servlet/cmd/UserDeleteCmd.java
@@ -9,6 +9,7 @@ import cz.svjis.bean.User;
 import cz.svjis.bean.UserDAO;
 import cz.svjis.servlet.CmdContext;
 import cz.svjis.servlet.Command;
+import cz.svjis.validator.Validator;
 import javax.servlet.RequestDispatcher;
 
 /**
@@ -23,15 +24,34 @@ public class UserDeleteCmd extends Command {
 
     @Override
     public void execute() throws Exception {
+        
+        String parId = getRequest().getParameter("id");
+        
+        if (!validateInput(parId)) {
+            RequestDispatcher rd = getRequest().getRequestDispatcher("/InputValidationError.jsp");
+            rd.forward(getRequest(), getResponse());
+            return;
+        }
+        
         UserDAO userDao = new UserDAO(getCnn());
 
         User u = new User();
-        u.setId(Integer.valueOf(getRequest().getParameter("id")));
+        u.setId(Integer.valueOf(parId));
         u.setCompanyId(getCompany().getId());
         userDao.deleteUser(u);
         String url = "Dispatcher?page=userList";
         getRequest().setAttribute("url", url);
         RequestDispatcher rd = getRequest().getRequestDispatcher("/_refresh.jsp");
         rd.forward(getRequest(), getResponse());
+    }
+    
+    private boolean validateInput(String parId) {
+        boolean result = true;
+        
+        if (!Validator.validateInteger(parId, 0, Validator.maxIntAllowed)) {
+            result = false;
+        }
+
+        return result;
     }
 }
